Take a Ten Minutes Walk
created empty arrays to hold the new values.
Mapped over the values of the array and if any of the values matched "n", "s", "e", "w" - then we would push 1 to the array.

we check if the length of n and s match - then we check if w and e match and then we check if the walk.length is equal to 10

All tests passed!